Added InitializeCustomBoneMapping function to the AnimInstance to init custom bone mappings
Made the custom controller class not load if the official valve controller is active
Also made the custom controller class not load if the beta input branch is not being used (CVAR)

Removed my temp custom CVAR

Now using the same skeletal action name as the official valve plugin, will have to re-bind in the input mappings.

Very rough and super expensive replication added to the skeletons meshes, need to work on it a lot, only for initial testing.